Move command loop to mqtt agent file
